Fix tests
This change is needed because the QP is now having an initial value
and in places where it is not set it is bound to first render
with the initial params and then whit the initial params in the
query params. This should however happen only on the first render
and the performance impact is small.